refactor: 아호코라식으로 교체(#103)
